feature/auth-01 - Auth

+++ product/src/app.controller.ts
@@ -1,4 +1,4 @@
-import { Controller, Get } from '@nestjs/common';
+import { Controller, Get, UseGuards } from '@nestjs/common';
 import { AppService } from './app.service';
 import {
   GetRequest,
@@ -12,31 +12,38 @@ import {
   PRODUCT_CR_UD_SERVICE_NAME,
   AddRequest,
   AddResponse,
-  ProductCRUDServiceController
-} from "./stubs/product/v1alpha/product";
-import {Metadata} from "@grpc/grpc-js";
-import {GrpcMethod} from "@nestjs/microservices";
+  ProductCRUDServiceController,
+} from './stubs/product/v1alpha/product';
+import { Metadata } from '@grpc/grpc-js';
+import { GrpcMethod, RpcException } from '@nestjs/microservices';
+import { UserCRUDServiceClient } from './stubs/user/v1alpha/user';
+import { UserService } from './user/user.service';
+import { JwtService } from '@nestjs/jwt';
 
 @Controller()
 @ProductCRUDServiceControllerMethods()
 export class AppController implements ProductCRUDServiceController {
-  constructor(private readonly appService: AppService) {}
+  constructor(
+    private readonly appService: AppService,
+    private userService: UserService,
+    private jwtService: JwtService,
+  ) {}
   async get(request: GetRequest, metadata?: Metadata): Promise<GetResponse> {
     let product: Product;
     let products: Product[] = [];
 
-    if(request.id){
+    if (request.id) {
       product = await this.appService.findById(request.id);
-      return {products: [product]};
-    }else{
+      return { products: [product] };
+    } else {
       products = await this.appService.findAll();
-      return {products}
+      return { products };
     }
   }
-    async update(
-        request: UpdateRequest,
-        metadata?: Metadata
-    ): Promise<UpdateResponse> {
+  async update(
+    request: UpdateRequest,
+    metadata?: Metadata,
+  ): Promise<UpdateResponse> {
     const id = request.id;
 
     Object.keys(request).forEach(
@@ -45,26 +52,95 @@ export class AppController implements ProductCRUDServiceController {
 
     delete request.id;
 
-    const product = await this.appService.update(id, request);
+    if (!request.userId) {
+      throw new RpcException('No user provided');
+    }
+
+    const userExist = await this.isUserExist(request.userId);
+
+    if (!userExist) {
+      throw new RpcException("User doesn't exist");
+    }
+
+    await this.isUserOwner(request.userId, id, 'update');
+
+    const product = await this.appService.update(id, {
+      name: request.name,
+      description: request.description,
+      price: request.price,
+    });
 
     return { product };
+  }
 
+  async delete(
+    request: DeleteRequest,
+    metadata?: Metadata,
+  ): Promise<DeleteResponse> {
+    if (!request.userId) {
+      throw new RpcException('No user provided');
     }
 
-    async delete(
-        request: DeleteRequest,
-        metadata?: Metadata
-    ): Promise<DeleteResponse> {
-        const product = await this.appService.delete(request.id);
+    const userExist = await this.isUserExist(request.userId);
 
-        return { product };
+    if (!userExist) {
+      throw new RpcException("User doesn't exist");
     }
 
-    @GrpcMethod(PRODUCT_CR_UD_SERVICE_NAME)
-    async add(request: AddRequest): Promise<AddResponse> {
-        const product = await this.appService.create(request as any);
+    await this.isUserOwner(request.userId, request.id, 'delete');
 
-        return { product };
+    const product = await this.appService.delete(request.id);
+
+    return { product };
+  }
+
+  @GrpcMethod(PRODUCT_CR_UD_SERVICE_NAME)
+  async add(request: AddRequest): Promise<AddResponse> {
+    if (!request.userId) {
+      throw new RpcException('No user provided');
     }
 
+    const userExist = await this.isUserExist(request.userId);
+
+    if (!userExist) {
+      throw new RpcException("User doesn't exist");
+    }
+
+    const product = await this.appService.create(request as any);
+    return { product };
+  }
+
+  async isUserOwner(userId, productId, action) {
+    const product = await this.appService.findById(productId);
+
+    if (userId !== product.userId) {
+      if (action === 'update') {
+        throw new RpcException(
+          "You can't edit this product because you aren't the owner",
+        );
+      } else if (action === 'delete') {
+        throw new RpcException(
+          "You can't delete this product because you aren't the owner",
+        );
+      }
+    }
+  }
+
+  async isUserExist(userId) {
+    const userExist = await this.userService.getUser(
+      {
+        id: userId,
+        firstname: undefined,
+        lastname: undefined,
+        email: undefined,
+      },
+      { Authorization: `Bearer ${this.jwtService.sign({ internal: true })}` },
+    );
+
+    if (!userExist) {
+      return false;
+    }
+
+    return true;
+  }
 }
